Extract validator expression argument helpers
